Combine Token and Tokenizer, remove unnecessary usings
