+++ MessageGeneration/MessageParser.cs
@@ -1,8 +1,4 @@
-﻿using System.Globalization;
-using System.IO;
-
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
+﻿using Microsoft.CodeAnalysis.CSharp;
 
 namespace RosNet.MessageGeneration;
